refactor(rup): charts niño sano no es un elemento-rup
