+++ src/app/modules/rup/components/ejecucion/resumen-paciente/chart.html
